+++ WealthManagementAssessment/Domain/Contracts/Interfaces/IStockService.cs
@@ -5,6 +5,6 @@ namespace WealthManagementAssessment.Domain.Contracts.Interfaces
     public interface IStockService
     {
         //usar DECIMAL em tudo!!
-        double StockEngine(List<Investment> investments, DateTime valuationDate);
+        decimal StockEngine(List<Investment> investments, DateTime valuationDate);
     }
 }
